Don't rely on change events
Debounces `input` events without relying solely on the `change` event.

The `change` event isn't fired in a consistent manner. The expected behavior is for it to fire only on the trailing edge of the debounce, but sometimes it fires only on the leading edge. When this happens, the change will never be communicated to the handlers until another change happens.