Fix several minor style issues caused by change to interface skeleton (#90)
* Set a max width for the interface skeleton

This helps prevent it going too wide depending on the content

* Restore behaviour of fullscreen mode on html node

It previously added is-fullscreen-mode to the html node. Until there is a consistent pattern, let's restore that

* Explicitly close the inserter

Rather than rely on focus being changed and then closing the inserter, explicitly close it. This is a bit faster, and fixes the situation where opening the pattern explorer means that the block inserter can't be closed.

* Don't apply button style when active

This allows active styles to work

* Bundle

* Update changelog